GPII-20: First steps at updating linux acceptance testing to work with framework changes in universal
