--- tests/integrationTests/setup1/configs/development-config.json
@@ -1,23 +1,0 @@
-{
-    "typeName": "flowManager.preferencesServer.solutionsRegistry.deviceReporter.matchMaker.ontologyServer.development",
-    "options": {
-        "gradeNames": ["autoInit", "fluid.littleComponent"],
-        "components": {
-            "server": {
-                "type": "kettle.server",
-                "options": {
-                    "logging": true,
-                    "port": 8081
-                }
-            }
-        }
-    },
-    "includes": [
-        "../../../../../node_modules/universal/gpii/node_modules/deviceReporter/configs/development.json",
-        "../../../../../node_modules/universal/gpii/node_modules/flowManager/configs/development.json",
-        "../../../../../node_modules/universal/gpii/node_modules/matchMaker/configs/development.json",
-        "../../../../../node_modules/universal/gpii/node_modules/ontologyServer/configs/development.json",
-        "../../../../../node_modules/universal/gpii/node_modules/preferencesServer/configs/development.json",
-        "../../../../../node_modules/universal/gpii/node_modules/solutionsRegistry/configs/development.json"
-    ]
-}
